cs/Language/LexicalStructure.cs
//namespace Language;
public class LexicalStructure
{
    public static string UsingUnicodeEscapeSequences(bool \u0066)
    {
        char c = '\u0066';
        if (\u0066)
        {
            return c.ToString();
        }

        return "";
    }
    
    public static string UsingNormalCharacters(bool f)
    {
        char c = 'f';
        if (f)
        {
            return c.ToString();
        }

        return "";
    }

    public static bool @static(bool @bool)
    {
        return @bool;
    }

}
